folder naming changed and types are moved
