+++ src/utils/types/board.ts
@@ -1,0 +1,5 @@
+import type { BoardItem } from "./dashboard";
+
+export type BoardsState = {
+  items: BoardItem[];
+};
